Added ability to loop animation.
